refactor(runtime): split server builders and lifecycle handlers
