+++ rust/crates/runtime/src/matrix_cloud_runtime.rs
@@ -436,82 +436,6 @@ mod tests {
         unsafe { std::env::remove_var("MATRIXONE_PASSWORD") };
     }
 
-    #[test]
-    fn matrix_runtime_tracks_session_sync_tasks() {
-        let source = include_str!("matrix_cloud_runtime.rs");
-        assert!(
-            source.contains("session_sync_tasks: Mutex<JoinSet<()>>"),
-            "MatrixCloudRuntime should keep a tracked JoinSet for session sync sidecars"
-        );
-        assert!(
-            source.contains("pub fn spawn_session_sync_task"),
-            "MatrixCloudRuntime should expose a helper for tracked session sync spawns"
-        );
-        assert!(
-            source.contains("tasks.try_join_next()"),
-            "spawn_session_sync_task should opportunistically reap completed sync tasks"
-        );
-    }
-
-    #[test]
-    fn shutdown_waits_for_session_sync_tasks() {
-        let source = include_str!("matrix_cloud_runtime.rs");
-        assert!(
-            source.contains("std::mem::take(&mut *tasks)"),
-            "shutdown_ingestion_and_wait should drain tracked session sync tasks"
-        );
-        assert!(
-            source.contains("session_sync_tasks.join_next().await"),
-            "shutdown_ingestion_and_wait should await tracked session sync tasks before exit"
-        );
-    }
-
-    #[test]
-    fn serve_calls_shutdown_ingestion_and_wait() {
-        let source = include_str!("server/mod.rs");
-        assert!(
-            source.contains("shutdown_ingestion_and_wait"),
-            "serve() must call shutdown_ingestion_and_wait after axum serve returns"
-        );
-    }
-
-    #[test]
-    fn spawn_data_cleanup_respects_cancellation_token() {
-        let source = include_str!("server/mod.rs");
-        assert!(
-            source.contains("CancellationToken"),
-            "spawn_data_cleanup must accept a CancellationToken so shutdown can drain it"
-        );
-        assert!(
-            source.contains("cancel.cancelled()") || source.contains("cancel_token.cancelled()"),
-            "spawn_data_cleanup loop must select! on the cancel token"
-        );
-    }
-
-    /// HIGH #4: BridgePersistTracker trait must exist and be object-safe.
-    #[test]
-    fn bridge_persist_tracker_trait_is_defined() {
-        let source = include_str!("matrix_cloud_runtime.rs");
-        assert!(
-            source.contains("pub trait BridgePersistTracker"),
-            "BridgePersistTracker trait must be declared in matrix_cloud_runtime"
-        );
-        assert!(
-            source.contains("fn track_persist_task"),
-            "BridgePersistTracker must expose track_persist_task"
-        );
-    }
-
-    /// HIGH #4: MatrixCloudRuntime must implement BridgePersistTracker.
-    #[test]
-    fn matrix_cloud_runtime_impls_bridge_persist_tracker() {
-        let source = include_str!("matrix_cloud_runtime.rs");
-        assert!(
-            source.contains("impl BridgePersistTracker for MatrixCloudRuntime"),
-            "MatrixCloudRuntime must implement BridgePersistTracker"
-        );
-    }
-
     /// HIGH #4: BridgePersistTracker functional test — future runs via a minimal test impl.
     #[tokio::test]
     async fn bridge_persist_tracker_future_runs_and_drains() {
@@ -540,66 +464,6 @@ mod tests {
         assert!(
             task_ran.load(Ordering::SeqCst),
             "tracked future must execute"
-        );
-    }
-
-    /// HIGH #4: InProcessChatTurnBridge must have a persist_tracker field.
-    #[test]
-    fn bridge_inprocess_has_persist_tracker_field() {
-        let source = include_str!("turn/bridge_inprocess.rs");
-        assert!(
-            source.contains("persist_tracker"),
-            "InProcessChatTurnBridge must have a persist_tracker field for HIGH #4"
-        );
-        assert!(
-            source.contains("BridgePersistTracker"),
-            "bridge_inprocess.rs must reference BridgePersistTracker"
-        );
-    }
-
-    /// HIGH #4: The fire-and-forget persist paths in bridge_inprocess.rs must be
-    /// replaced with the tracked path (no raw TODO(audit-#3) deferred comment).
-    #[test]
-    fn bridge_persist_uses_tracker_not_raw_spawn() {
-        let source = include_str!("turn/bridge_inprocess.rs");
-        assert!(
-            !source.contains("TODO(audit-#3)"),
-            "audit-#3 TODO must be resolved — persist tasks should be tracked now"
-        );
-    }
-
-    #[test]
-    fn with_encryptor_stores_encryptor() {
-        let source = include_str!("matrix_cloud_runtime.rs");
-        assert!(
-            source.contains("encryptor: Option<Arc<astra_services::FernetTokenEncryptor>>"),
-            "MatrixCloudRuntime must store optional encryptor for model resolution"
-        );
-        assert!(
-            source.contains("fn with_encryptor"),
-            "MatrixCloudRuntime must expose with_encryptor builder method"
-        );
-        assert!(
-            source.contains("fn resolve_memory_model"),
-            "MatrixCloudRuntime must expose resolve_memory_model for memory relevance"
-        );
-    }
-
-    #[test]
-    fn resolve_memory_model_requires_encryptor() {
-        let source = include_str!("matrix_cloud_runtime.rs");
-        assert!(
-            source.contains("let enc = self.encryptor.as_ref()?;"),
-            "resolve_memory_model must early-return None when encryptor is missing"
-        );
-    }
-
-    #[test]
-    fn resolve_memory_model_returns_llm_conn_params() {
-        let source = include_str!("matrix_cloud_runtime.rs");
-        assert!(
-            source.contains("crate::memory_relevance::LlmConnParams"),
-            "resolve_memory_model must return LlmConnParams from memory_relevance module"
         );
     }
 }
