+++ rust/crates/runtime/src/server/router_builder/catalog.rs
@@ -1,0 +1,128 @@
+use super::*;
+
+pub(super) fn add_routes(router: Router<AppState>) -> Router<AppState> {
+    router
+        .route(
+            "/models",
+            post(models::create_model_handler).get(models::list_models_handler),
+        )
+        .route("/models/memory", get(models::get_memory_model_handler))
+        .route(
+            "/models/{model_name}",
+            get(models::get_model_handler)
+                .put(models::update_model_handler)
+                .delete(models::delete_model_handler),
+        )
+        .route(
+            "/models/{model_name}/check",
+            post(models::check_model_handler),
+        )
+        .route(
+            "/skills",
+            post(skills::register_skill_handler).get(skills::list_skills_handler),
+        )
+        .route("/skills/status", get(skills::get_skill_status_handler))
+        .route("/skills/publish", post(skills::publish_skill_handler))
+        .route(
+            "/skills/user",
+            get(user_skill_handlers::list_user_skills_handler)
+                .post(user_skill_handlers::create_user_skill_handler),
+        )
+        .route(
+            "/skills/user/{skill_name}/versions",
+            get(user_skill_handlers::list_user_skill_versions_handler)
+                .post(user_skill_handlers::submit_user_skill_version_handler),
+        )
+        .route(
+            "/skills/user/{skill_name}/evaluations",
+            post(user_skill_handlers::record_user_skill_evaluation_handler),
+        )
+        .route(
+            "/skills/user/{skill_name}/activate",
+            post(user_skill_handlers::activate_user_skill_handler),
+        )
+        .route(
+            "/skills/user/{skill_name}/install",
+            post(user_skill_handlers::install_user_skill_handler),
+        )
+        .route(
+            "/skills/{skill_name}/info",
+            get(skills::get_skill_info_handler),
+        )
+        .route("/skills/{skill_id}", get(skills::get_skill_handler))
+        .route(
+            "/skills/{skill_id}/versions",
+            get(skills::list_skill_versions_handler),
+        )
+        .route(
+            "/skills/{skill_name}/unpublish",
+            post(skills::unpublish_skill_handler),
+        )
+        .route(
+            "/skills/{skill_name}/config/validate",
+            get(skill_config::validate_config_handler),
+        )
+        .route(
+            "/skills/{skill_name}/config",
+            get(skill_config::get_effective_config_handler),
+        )
+        .route(
+            "/skills/{skill_name}/config/{setting_name}",
+            axum::routing::put(skill_config::set_setting_handler)
+                .delete(skill_config::delete_setting_handler),
+        )
+        .route(
+            "/skills/{skill_name}/resources",
+            get(skill_config::list_resources_handler),
+        )
+        .route(
+            "/skills/{skill_name}/resources/{resource_key}",
+            axum::routing::put(skill_config::bind_resource_handler)
+                .delete(skill_config::unbind_resource_handler),
+        )
+        .route(
+            "/marketplace/install",
+            post(marketplace::install_skill_handler),
+        )
+        .route(
+            "/marketplace/uninstall",
+            post(marketplace::uninstall_skill_handler),
+        )
+        .route(
+            "/marketplace/upgrade",
+            post(marketplace::upgrade_skill_handler),
+        )
+        .route(
+            "/marketplace/rollback",
+            post(marketplace::rollback_skill_handler),
+        )
+        .route(
+            "/marketplace/installed",
+            get(marketplace::list_installed_handler),
+        )
+        .route(
+            "/marketplace/credentials",
+            post(marketplace::save_credential_handler)
+                .delete(marketplace::delete_credential_handler),
+        )
+        .route(
+            "/marketplace/skills/{skill_name}/publish",
+            post(marketplace::publish_skill_handler),
+        )
+        .route(
+            "/marketplace/skills/{skill_name}/deprecate",
+            post(marketplace::deprecate_skill_handler),
+        )
+        .route(
+            "/marketplace/quality-report",
+            post(marketplace::submit_quality_report_handler),
+        )
+        .route(
+            "/marketplace/stats/{skill_name}",
+            get(marketplace::get_skill_stats_handler),
+        )
+        .route(
+            "/marketplace/search",
+            get(marketplace::search_marketplace_handler),
+        )
+}
