+++ rust/crates/runtime/src/server/router_builder/operations/insights.rs
@@ -1,0 +1,97 @@
+use super::super::*;
+
+pub(super) fn add_routes(router: Router<AppState>) -> Router<AppState> {
+    router
+        .route(
+            "/evaluation/quality/trend",
+            get(evaluation::quality_trend_handler),
+        )
+        .route("/evaluation/drift", get(evaluation::drift_handler))
+        .route("/evaluation/gates", get(evaluation::gate_history_handler))
+        .route(
+            "/evaluation/calibration",
+            get(evaluation::calibration_handler),
+        )
+        .route(
+            "/evaluation/sessions/scores",
+            get(evaluation::session_scores_handler),
+        )
+        .route(
+            "/evaluation/gate/validate",
+            post(evaluation::gate_validate_handler),
+        )
+        .route("/evaluation/drift/run", post(evaluation::drift_run_handler))
+        .route("/evaluation/loop", post(evaluation::closed_loop_handler))
+        .route(
+            "/evaluation/trust-report",
+            get(evaluation::trust_report_handler),
+        )
+        .route(
+            "/evaluation/slo/dashboard",
+            get(evaluation::slo_dashboard_handler),
+        )
+        .route(
+            "/evaluation/slo/{agent_id}/history",
+            get(evaluation::slo_history_handler),
+        )
+        .route(
+            "/evaluation/observability/metrics",
+            get(evaluation::observability_metrics_handler),
+        )
+        .route(
+            "/evaluation/memory-health",
+            get(evaluation::memory_health_handler),
+        )
+        .route(
+            "/evaluation/memory-metrics",
+            get(evaluation::memory_metrics_handler),
+        )
+        .route(
+            "/evaluation/training-data/extract",
+            post(evaluation::training_data_extract_handler),
+        )
+        .route(
+            "/evaluation/training-data/{dataset_id}/export",
+            get(evaluation::training_data_export_handler),
+        )
+        .route(
+            "/introspection/memory",
+            get(introspection::get_memory_introspection_handler),
+        )
+        .route(
+            "/introspection/decision-trace",
+            get(introspection::get_decision_trace_handler),
+        )
+        .route(
+            "/introspection/tool-history",
+            get(introspection::get_tool_history_handler),
+        )
+        .route(
+            "/introspection/drift-check",
+            get(introspection::get_drift_check_handler),
+        )
+        .route(
+            "/introspection/skills",
+            get(introspection::get_skills_introspection_handler),
+        )
+        .route(
+            "/introspection/context/trend",
+            get(introspection::get_context_trend_handler),
+        )
+        .route(
+            "/introspection/context/snapshot",
+            get(introspection::get_context_snapshot_handler),
+        )
+        .route(
+            "/introspection/context/retrieval_quality",
+            get(introspection::get_retrieval_quality_handler),
+        )
+        .route(
+            "/introspection/memory/recall",
+            get(introspection::get_memory_recall_handler),
+        )
+        .route(
+            "/platform/snapshot",
+            get(platform_handlers::platform_snapshot_handler),
+        )
+}
