+++ rust/crates/runtime/src/server/router_builder/operations/work.rs
@@ -1,0 +1,94 @@
+use super::super::*;
+
+pub(super) fn add_routes(router: Router<AppState>) -> Router<AppState> {
+    router
+        .route("/runs", get(run_handlers::list_runs_handler))
+        .route(
+            "/tasks",
+            get(task_handlers::list_tasks_handler).post(task_handlers::create_task_handler),
+        )
+        .route("/tasks/{task_id}", get(task_handlers::get_task_handler))
+        .route("/tasks:rpc", post(task_handlers::task_rpc_handler))
+        .route(
+            "/preferences",
+            get(preferences_handlers::list_preferences_handler),
+        )
+        .route(
+            "/preferences/{key}",
+            axum::routing::put(preferences_handlers::put_preference_handler),
+        )
+        .route(
+            "/tasks/{task_id}/lease/claim",
+            post(task_handlers::post_task_lease_claim_handler),
+        )
+        .route(
+            "/tasks/{task_id}/lease/release",
+            post(task_handlers::post_task_lease_release_handler),
+        )
+        .route(
+            "/tasks/{task_id}/lease/renew",
+            post(task_handlers::post_task_lease_renew_handler),
+        )
+        .route(
+            "/tasks/{task_id}/lease",
+            get(task_handlers::get_task_lease_handler),
+        )
+        .route(
+            "/tasks/{task_id}/progress",
+            get(task_handlers::task_progress_handler),
+        )
+        .route(
+            "/tasks/{task_id}/status",
+            axum::routing::put(task_handlers::update_task_status_handler),
+        )
+        .route(
+            "/resources/usage",
+            get(resource_handlers::get_resource_usage_handler),
+        )
+        .route(
+            "/resources/limits",
+            get(resource_handlers::get_resource_limits_handler),
+        )
+        .route(
+            "/plans",
+            post(plan_handlers::create_plan_handler).get(plan_handlers::list_plans_handler),
+        )
+        .route(
+            "/plans/{plan_id}",
+            get(plan_handlers::get_plan_handler)
+                .put(plan_handlers::update_plan_handler)
+                .delete(plan_handlers::delete_plan_handler),
+        )
+        .route(
+            "/plans/{plan_id}/status",
+            get(plan_handlers::plan_status_handler),
+        )
+        .route(
+            "/plans/{plan_id}/execute",
+            post(plan_handlers::execute_plan_handler),
+        )
+        .route(
+            "/plans/{plan_id}/exit-plan-mode",
+            post(plan_handlers::exit_plan_mode_handler),
+        )
+        .route(
+            "/plans/{plan_id}/rewind",
+            post(plan_handlers::rewind_plan_handler),
+        )
+        .route(
+            "/plans/{plan_id}/redo-step",
+            post(plan_handlers::redo_step_handler),
+        )
+        .route(
+            "/plans/{plan_id}/step-runs",
+            get(plan_handlers::list_step_runs_handler).post(plan_handlers::start_step_run_handler),
+        )
+        .route(
+            "/plans/{plan_id}/step-runs/completed",
+            post(plan_handlers::post_completed_step_run_handler),
+        )
+        .route(
+            "/plans/{plan_id}/step-runs/{run_id}/finish",
+            post(plan_handlers::finish_step_run_handler),
+        )
+}
