+++ rust/crates/runtime/src/server/state_builder/runtime.rs
@@ -1,0 +1,197 @@
+use super::*;
+
+pub(super) async fn build_runtime_wiring(
+    settings: &AppSettings,
+    shared_pool: &SharedPool,
+    lease_hold_cache: &Arc<TaskLeaseHoldCache>,
+    run_encryptor: &Arc<FernetTokenEncryptor>,
+    state: &AppState,
+) -> Result<RuntimeWiring, Box<dyn std::error::Error>> {
+    let run_store = Arc::new(astra_services::runs::DatabaseRunStateStore::new(
+        shared_pool.clone(),
+    ));
+    let state_projection_store = Arc::new(astra_services::DatabaseStateProjectionStore::new(
+        shared_pool.clone(),
+    ));
+    let run_engine = crate::server::run_engine::RunEngine::new(run_store)
+        .with_projection_store(Arc::clone(&state_projection_store));
+    recover_active_runs(&run_engine).await;
+
+    let profile_registry = Arc::new(default_agent_profile_registry());
+    let progress_broadcaster = Arc::new(crate::orchestration::ProgressBroadcaster::default());
+    let delegation_tracker = Arc::new(
+        crate::server::delegation_engine::DelegationTracker::new()
+            .with_progress_broadcaster(Arc::clone(&progress_broadcaster)),
+    );
+    let user_id = astra_core::cli_user_id();
+    let matrix_rt = Arc::new(
+        crate::matrix_cloud_runtime::MatrixCloudRuntime::attach(
+            shared_pool.clone(),
+            "default",
+            &user_id,
+            Arc::clone(lease_hold_cache),
+        )
+        .with_encryptor(Arc::clone(run_encryptor)),
+    );
+    let memory_extraction_service = matrix_rt.clone_memory_extraction_service();
+
+    let sub_run_executor: Arc<dyn crate::server::delegation_engine::SubRunExecutor> = {
+        let mut exec = super::run_lifecycle::ServerSubRunExecutor::new(
+            settings.matrixone.clone(),
+            Arc::clone(run_encryptor),
+            state.edge_callback_ledger.clone(),
+        )
+        .with_pool(shared_pool.clone())
+        .with_edge_connection_pool(state.edge_connection_pool.clone())
+        .with_skill_service(state.skill_service.clone());
+        if let Some(svc) = memory_extraction_service.as_ref() {
+            exec = exec.with_memory_extraction_service(Arc::clone(svc));
+        }
+        Arc::new(exec)
+    };
+    let delegation_engine = Arc::new(
+        crate::server::delegation_engine::DelegationEngine::with_executor(
+            Arc::new(tokio::sync::RwLock::new((*profile_registry).clone())),
+            Arc::new(run_engine.clone()),
+            delegation_tracker,
+            sub_run_executor,
+        )
+        .with_projection_store(Arc::clone(&state_projection_store)),
+    );
+
+    let resource_governor = initialize_resource_governor(shared_pool).await;
+    let mut run_lifecycle = super::run_lifecycle::AgenticRunLifecycleService::new(
+        settings.matrixone.clone(),
+        Arc::clone(run_encryptor),
+        state.edge_callback_ledger.clone(),
+    )
+    .with_pool(shared_pool.clone())
+    .with_run_engine(run_engine)
+    .with_delegation_engine(Arc::clone(&delegation_engine))
+    .with_edge_connection_pool(state.edge_connection_pool.clone())
+    .with_resource_governor(resource_governor.clone())
+    .with_skill_service(state.skill_service.clone())
+    .with_hook_db_writer(state.turn_persistence.hook_db_writer.clone())
+    .with_observer_worker(state.turn_persistence.observer_worker.clone())
+    .with_tool_event_writer(state.turn_persistence.tool_event_writer.clone())
+    .with_auxiliary_event_writer(state.turn_persistence.auxiliary_event_writer.clone());
+    if let Some(svc) = memory_extraction_service.as_ref() {
+        run_lifecycle = run_lifecycle.with_memory_extraction_service(Arc::clone(svc));
+    }
+
+    #[cfg(feature = "harness")]
+    let run_lifecycle = run_lifecycle.with_harness_registry(state.harness_registry.clone());
+
+    let team_store = initialize_team_store(shared_pool, &user_id).await;
+    Ok(RuntimeWiring {
+        matrix_rt,
+        run_lifecycle,
+        profile_registry,
+        delegation_engine,
+        team_store,
+        resource_governor,
+    })
+}
+
+async fn recover_active_runs(run_engine: &crate::server::run_engine::RunEngine) {
+    match run_engine.recover_active_runs().await {
+        Ok(recovered_runs) => {
+            if !recovered_runs.is_empty() {
+                let waiting = recovered_runs
+                    .iter()
+                    .filter(|run| run.status == astra_core::STATUS_WAITING)
+                    .count();
+                let failed = recovered_runs
+                    .iter()
+                    .filter(|run| run.status == astra_core::STATUS_FAILED)
+                    .count();
+                tracing::warn!(
+                    target: "astra_runtime::state_builder",
+                    recovered_total = recovered_runs.len(),
+                    recovered_waiting = waiting,
+                    recovered_failed = failed,
+                    "recovered durable active runs during startup"
+                );
+            }
+        }
+        Err(error) => {
+            tracing::error!(
+                target: "astra_runtime::state_builder",
+                error = %error,
+                "failed to recover durable active runs during startup"
+            );
+        }
+    }
+}
+
+async fn initialize_resource_governor(
+    shared_pool: &SharedPool,
+) -> std::sync::Arc<dyn astra_services::resource_governor::ResourceGovernor> {
+    let resource_governor =
+        astra_services::resource_governor::DatabaseResourceGovernor::new(shared_pool.clone());
+    if let Err(error) = resource_governor.ensure_tables().await {
+        tracing::warn!(
+            target: "astra_runtime::state_builder",
+            error = %error,
+            "resource_governor table init failed"
+        );
+    }
+    std::sync::Arc::new(resource_governor)
+}
+
+async fn initialize_team_store(
+    shared_pool: &SharedPool,
+    user_id: &str,
+) -> Arc<dyn astra_services::team_persistence::TeamPersistenceService> {
+    let team_store =
+        astra_services::team_persistence::MatrixOneTeamStore::new(shared_pool.get().clone());
+    if let Err(error) = team_store.ensure_builtins(user_id).await {
+        tracing::warn!(
+            target: "astra_runtime::state_builder",
+            error = %error,
+            user_id = %user_id,
+            "team builtins seed failed"
+        );
+    }
+    Arc::new(team_store)
+}
+
+pub(super) fn default_agent_profile_registry() -> astra_services::AgentProfileRegistry {
+    use astra_services::coordination::{AgentProfile, AgentTier};
+
+    let mut profile_registry = astra_services::AgentProfileRegistry::new();
+
+    let mut orch = AgentProfile::new("orchestrator", "Orchestrator", AgentTier::Orchestrator);
+    orch.system_prompt = Some(
+        "You are the orchestrator agent. Coordinate sub-agents to complete complex tasks."
+            .to_string(),
+    );
+    let _ = profile_registry.register(orch);
+
+    let mut coder = AgentProfile::new("coder", "Coder", AgentTier::System);
+    coder.system_prompt =
+        Some("You are a coding agent. Write, edit, and debug code to complete tasks.".to_string());
+    coder.skill_filter = vec![
+        "bash".into(),
+        "read_file".into(),
+        "write_file".into(),
+        "str_replace".into(),
+        "git_commit".into(),
+    ];
+    let _ = profile_registry.register(coder);
+
+    let mut reviewer = AgentProfile::new("reviewer", "Reviewer", AgentTier::System);
+    reviewer.system_prompt = Some(
+        "You are a code review agent. Review code for bugs, security, and best practices."
+            .to_string(),
+    );
+    reviewer.skill_filter = vec!["read_file".into(), "bash".into()];
+    let _ = profile_registry.register(reviewer);
+
+    let mut writer = AgentProfile::new("writer", "Writer", AgentTier::User);
+    writer.system_prompt =
+        Some("You are a documentation writer. Create clear, concise docs.".to_string());
+    let _ = profile_registry.register(writer);
+
+    profile_registry
+}
